+++ plugins/core/Standard/qRANSAC_SD/extern/RANSAC_SD/ConePrimitiveShapeConstructor.cpp
@@ -4,6 +4,12 @@
 #include "ScoreComputer.h"
 #include <GfxTL/NullClass.h>
 
+ConePrimitiveShapeConstructor::ConePrimitiveShapeConstructor(float maxConeRadius, float maxAngleRadians, float maxConeLength)
+	:	m_maxConeLength(maxConeLength),
+	m_maxConeRadius(maxConeRadius),
+	m_maxAngle(maxAngleRadians)
+{}
+
 size_t ConePrimitiveShapeConstructor::Identifier() const
 {
 	return 3;
@@ -19,13 +25,28 @@ PrimitiveShape *ConePrimitiveShapeConstructor::Construct(
 	const MiscLib::Vector< Vec3f > &normals) const
 {
 	Cone cone;
-	if(!cone.Init(points[0], points[1], points[2], normals[0], normals[1],
+	if (!cone.Init(points[0], points[1], points[2], normals[0], normals[1],
 		normals[2]))
+	{
 		return NULL;
-	if(cone.Angle() > 1.4835298641951801403851371532153)
-		// do not allow cones with an opening angle of more than 85 degrees
+	}
+	if (cone.Angle() > 1.4835298641951801403851371532153 || // do not allow cones with an opening angle of more than 85 degrees
+		cone.Angle() > m_maxAngle) 
+	{
 		return NULL;
-	return new ConePrimitiveShape(cone);
+	}
+	if (m_maxConeRadius < std::numeric_limits< float >::infinity() 
+		|| m_maxConeLength < std::numeric_limits< float >::infinity())
+	{
+		Cone::ConeInfo ci = cone.GetInfo(points);
+		if (ci.height > m_maxConeLength || ci.maxRadius > m_maxConeRadius || ci.minRadius > m_maxConeRadius)
+		{
+			return NULL;
+		}
+	}
+	
+
+	return new ConePrimitiveShape(cone, m_maxConeRadius, m_maxAngle, m_maxConeLength);
 }
 
 PrimitiveShape *ConePrimitiveShapeConstructor::Construct(
@@ -34,7 +55,22 @@ PrimitiveShape *ConePrimitiveShapeConstructor::Construct(
 	Cone cone;
 	if(!cone.Init(samples))
 		return NULL;
-	return new ConePrimitiveShape(cone);
+	if (cone.Angle() > 1.4835298641951801403851371532153 || // do not allow cones with an opening angle of more than 85 degrees
+		cone.Angle() > m_maxAngle)
+	{
+		return NULL;
+	}
+	if (m_maxConeRadius < std::numeric_limits< float >::infinity()
+		|| m_maxConeLength < std::numeric_limits< float >::infinity())
+	{
+		Cone::ConeInfo ci = cone.GetInfo(samples);
+		if (ci.height > m_maxConeLength || ci.maxRadius > m_maxConeRadius || ci.minRadius > m_maxConeRadius)
+		{
+			return NULL;
+		}
+	}
+
+	return new ConePrimitiveShape(cone, m_maxConeRadius, m_maxAngle, m_maxConeLength);
 }
 
 PrimitiveShape *ConePrimitiveShapeConstructor::Deserialize(std::istream *i,
@@ -42,7 +78,7 @@ PrimitiveShape *ConePrimitiveShapeConstructor::Deserialize(std::istream *i,
 {
 	Cone cone;
 	cone.Init(binary, i);
-	ConePrimitiveShape *shape = new ConePrimitiveShape(cone);
+	ConePrimitiveShape *shape = new ConePrimitiveShape(cone, m_maxConeRadius, m_maxAngle, m_maxConeLength);
 	return shape;
 }