+++ plugins/core/Standard/qRANSAC_SD/extern/RANSAC_SD/CylinderPrimitiveShapeConstructor.cpp
@@ -2,6 +2,14 @@
 #include "ScoreComputer.h"
 #include <GfxTL/NullClass.h>
 
+CylinderPrimitiveShapeConstructor::CylinderPrimitiveShapeConstructor(
+	float minCylinderRadius, float maxCylinderRadius, float maxCylinderLength)
+	: m_minCylinderRadius(minCylinderRadius)
+	, m_maxCylinderRadius(maxCylinderRadius)
+	, m_maxCylinderLength(maxCylinderLength)
+
+{}
+
 size_t CylinderPrimitiveShapeConstructor::Identifier() const
 {
 	return 2;
@@ -9,7 +17,9 @@ size_t CylinderPrimitiveShapeConstructor::Identifier() const
 
 unsigned int CylinderPrimitiveShapeConstructor::RequiredSamples() const
 {
-	return 2;
+	return 4; //2; Because the specific 2point/normal 
+			  //   constructor is not called in reality 
+			  //   we need 4 points
 }
 
 PrimitiveShape *CylinderPrimitiveShapeConstructor::Construct(
@@ -19,18 +29,30 @@ PrimitiveShape *CylinderPrimitiveShapeConstructor::Construct(
 	Cylinder cy;
 	MiscLib::Vector< Vec3f > samples(points);
 	std::copy(normals.begin(), normals.end(), std::back_inserter(samples));
-	if(!cy.Init(samples))
+	if (!cy.Init(samples))
+	{
 		return NULL;
-	return new CylinderPrimitiveShape(cy);
+	}
+	if (cy.Radius() > m_minCylinderRadius && cy.Radius() < m_maxCylinderRadius)
+	{
+		return new CylinderPrimitiveShape(cy, m_minCylinderRadius, m_maxCylinderRadius, m_maxCylinderLength);
+	}
+	return NULL;
 }
 
 PrimitiveShape *CylinderPrimitiveShapeConstructor::Construct(
 	const MiscLib::Vector< Vec3f > &samples) const
 {
 	Cylinder cy;
-	if(!cy.Init(samples))
+	if (!cy.Init(samples))
+	{
 		return NULL;
-	return new CylinderPrimitiveShape(cy);
+	}
+	if (cy.Radius() > m_minCylinderRadius && cy.Radius() < m_maxCylinderRadius)
+	{
+		return new CylinderPrimitiveShape(cy, m_minCylinderRadius, m_maxCylinderRadius, m_maxCylinderLength);
+	}
+	return NULL;
 }
 
 PrimitiveShape *CylinderPrimitiveShapeConstructor::Deserialize(