+++ plugins/core/Standard/qRANSAC_SD/extern/RANSAC_SD/SpherePrimitiveShape.cpp
@@ -28,14 +28,18 @@
 #include "PlanePrimitiveShape.h"
 extern MiscLib::performance_t totalTime_sphereConnected;
 
-SpherePrimitiveShape::SpherePrimitiveShape(const Sphere &s)
+SpherePrimitiveShape::SpherePrimitiveShape(const Sphere &s, float minRadius, float maxRadius)
 : m_sphere(s)
+, m_minRadius(minRadius)
+, m_maxRadius(maxRadius)
 , m_parametrization(m_sphere)
 {}
 
 SpherePrimitiveShape::SpherePrimitiveShape(const SpherePrimitiveShape &sps)
 : BitmapPrimitiveShape(sps)
 , m_sphere(sps.m_sphere)
+, m_minRadius(sps.m_minRadius)
+, m_maxRadius(sps.m_maxRadius)
 , m_parametrization(sps.m_parametrization)
 {
 	m_parametrization.Shape(m_sphere);
@@ -313,20 +317,38 @@ void SpherePrimitiveShape::SuggestSimplifications(const PointCloud &pc,
 				&samples[i * 5 + j], &samples[i * 5 + j + c]);
 	}
 
+	float d = 0;
+	float bestSum = 0;
+
+	for (size_t i = 0; i < c; ++i)
+	{
+		d = m_sphere.Distance(samples[i]);
+		bestSum += d;
+	}
+
 	Plane plane;
 	if(plane.LeastSquaresFit(samples.begin(), samples.begin() + c))
 	{
 		bool failed = false;
-		for(size_t i = 0; i < c; ++i)
-			if(plane.Distance(samples[i]) > distThresh)
+		float sum = 0;
+		for (size_t i = 0; i < c; ++i)
+		{
+			d = plane.Distance(samples[i]);
+			sum += d;
+			if (d > distThresh)
 			{
 				failed = true;
 				break;
 			}
-		if(!failed)
+		}
+		if (!failed)
 		{
-			suggestions->push_back(new PlanePrimitiveShape(plane));
-			suggestions->back()->Release();
+			if (sum < bestSum)
+			{
+				bestSum = sum;
+				suggestions->push_back(new PlanePrimitiveShape(plane));
+				suggestions->back()->Release();
+			}
 		}
 	}