+++ plugins/core/Standard/qRANSAC_SD/extern/RANSAC_SD/SpherePrimitiveShape.h
@@ -21,8 +21,8 @@ class DLL_LINKAGE SpherePrimitiveShape
 {
 public:
 	typedef LowStretchSphereParametrization ParametrizationType;
-	SpherePrimitiveShape() : m_parametrization(m_sphere) {}
-	SpherePrimitiveShape(const Sphere &s);
+	SpherePrimitiveShape() : m_parametrization(m_sphere), m_minRadius(-std::numeric_limits<float>::infinity()), m_maxRadius(std::numeric_limits<float>::infinity()) {}
+	SpherePrimitiveShape(const Sphere &s, float minRadius = -std::numeric_limits<float>::infinity(), float maxRadius = std::numeric_limits<float>::infinity());
 	SpherePrimitiveShape(const SpherePrimitiveShape &sps);
 	size_t Identifier() const;
 	unsigned int RequiredSamples() const { return Sphere::RequiredSamples; }
@@ -99,7 +99,16 @@ public:
 	bool InSpace(size_t u, size_t v, float epsilon,
 		const GfxTL::AABox< GfxTL::Vector2Df > &bbox, size_t uextent,
 		size_t vextent, Vec3f *p, Vec3f *n) const;
-
+	bool CheckGeneratedShapeWithinLimits(const PointCloud& pc,
+		MiscLib::Vector< size_t >::const_iterator begin,
+		MiscLib::Vector< size_t >::const_iterator end) override
+	{
+		if ((m_sphere.Radius() >= m_minRadius) && (m_sphere.Radius() <= m_maxRadius))
+		{
+			return true;
+		}
+		return false;
+	}
 private:
 	template< class IteratorT >
 	void ParametersImpl(IteratorT begin, IteratorT end,
@@ -107,6 +116,8 @@ private:
 
 private:
 	Sphere m_sphere;
+	float m_minRadius;
+	float m_maxRadius;
 	ParametrizationType m_parametrization;
 };