+++ plugins/core/Standard/qRANSAC_SD/extern/RANSAC_SD/TorusPrimitiveShapeConstructor.cpp
@@ -3,6 +3,19 @@
 #include "ScoreComputer.h"
 #include <GfxTL/NullClass.h>
 
+TorusPrimitiveShapeConstructor::TorusPrimitiveShapeConstructor(bool allowAppleShaped
+	, float minMinorRadius
+	, float minMajorRadius
+	, float maxMinorRadius
+	, float maxMajorRadius)
+	: m_allowAppleShaped(allowAppleShaped)
+	, m_minMinorRadius(minMinorRadius)
+	, m_minMajorRadius(minMajorRadius)
+	, m_maxMinorRadius(maxMinorRadius)
+	, m_maxMajorRadius(maxMajorRadius)
+{
+}
+
 size_t TorusPrimitiveShapeConstructor::Identifier() const
 {
 	return 4;
@@ -29,9 +42,20 @@ PrimitiveShape *TorusPrimitiveShapeConstructor::Construct(
 	const MiscLib::Vector< Vec3f > &samples) const
 {
 	Torus torus;
-	if(!torus.Init(samples))
+	if (!torus.Init(samples))
+	{
 		return NULL;
-	return new TorusPrimitiveShape(torus);
+	}
+	
+	if ((!m_allowAppleShaped && torus.IsAppleShaped()) ||
+		torus.MinorRadius() < m_minMinorRadius ||
+		torus.MajorRadius() < m_minMajorRadius ||
+		torus.MinorRadius() > m_maxMinorRadius ||
+		torus.MajorRadius() > m_maxMajorRadius)
+	{
+		return NULL;
+	}
+	return new TorusPrimitiveShape(torus, m_allowAppleShaped, m_maxMinorRadius, m_maxMajorRadius);
 }
 
 PrimitiveShape *TorusPrimitiveShapeConstructor::Deserialize(std::istream *i,