+++ plugins/core/Standard/qRANSAC_SD/extern/RANSAC_SD/TorusPrimitiveShapeConstructor.h
@@ -10,6 +10,11 @@ class DLL_LINKAGE TorusPrimitiveShapeConstructor
 : public PrimitiveShapeConstructor
 {
 public:
+	TorusPrimitiveShapeConstructor(bool allowAppleShaped = false
+		, float minMinorRadius = -std::numeric_limits<float>::infinity()
+		, float minMajorRadius = -std::numeric_limits<float>::infinity()
+		, float maxMinorRadius = std::numeric_limits<float>::infinity()
+		, float maxMajorRadius = std::numeric_limits<float>::infinity());
 	size_t Identifier() const;
 	unsigned int RequiredSamples() const;
 	PrimitiveShape *Construct(const MiscLib::Vector< Vec3f > &points,
@@ -17,6 +22,12 @@ public:
 	PrimitiveShape *Construct(const MiscLib::Vector< Vec3f > &samples) const;
 	PrimitiveShape *Deserialize(std::istream *i, bool binary = true) const;
 	size_t SerializedSize() const;
+private:
+	bool m_allowAppleShaped;
+	float m_minMinorRadius;
+	float m_minMajorRadius;
+	float m_maxMinorRadius;
+	float m_maxMajorRadius;
 };
 
 #endif