+++ CHANGELOG.md
@@ -4,10 +4,17 @@ CloudCompare Version History
 v2.12 (???) - (in development)
 ----------------------
 - Improvements
+    - RANSAC plugin
+        - Can save all leftover points into a new cloud (leftovers were points not assigned to a shape)
+        - Can select whether to use Least Squares fitting on found shapes (some shapes take a very long time to process this step specifically Cones)
+        - Can select whether to attemt to simplify shapes (torus->cone/cylinder/sphere/planes cone->cylinder/sphere/plane  cylinder->sphere/plane, sphere->plane)
+        - Can choose whether or not to have a random color assigned to each shape found.
+        - Ability to select min and max radii for various shapes (helps prevent giant spheres and cylinders from beating out the more likely plane feature)
     - Single Click Picking option added to display options menu
       - Single click picking can be disabled (can be very slow for very large point clouds) 
     - CommandLine mode new features
       - Added N_SIGMA_MIN and N_SIGMA_MAX options to the FILTER_SF command. Specify the option followed by a numeric value to filter by N * standardDeviation around the mean.
+
 	- Clipping box tool:
 		- former 'contours' renamed 'envelopes' for the sake of clarity
 		- ability to extract the real contours of the points inside each slice (single slice mode or 'repeat' mode)